requirements.txt
django
djangorestframework
drf-yasg
coreapi
django-cors-headers
